Another round of code cleanup
