--- realtime/.devcontainer/devcontainer.json
@@ -1,21 +1,0 @@
-{
-    "name": "NVQLink Development",
-    "image": "nvqlink-prototype:0.1.0",
-    "capAdd": ["SYS_PTRACE", "NET_ADMIN", "SYS_MODULE", "IPC_LOCK"],
-	"securityOpt": [ "seccomp=unconfined" ],
-    "mounts": [
-        "source=/sys/bus/pci/devices,target=/sys/bus/pci/devices,type=bind",
-        "source=/sys/kernel/mm/hugepages,target=/sys/kernel/mm/hugepages,type=bind",
-        "source=/dev,target=/dev,type=bind",
-        "source=/sys/devices,target=/sys/devices,type=bind",
-        "source=/opt/mellanox/gdrcopy,target=/opt/mellanox/gdrcopy,type=bind"
-    ],
-    "overrideCommand": true,
-    "remoteUser": "root",
-    "runArgs": [
-        "--privileged",
-        "--gpus=all",
-        "--network=host",
-        "--device=/dev/gdrdrv"
-    ]
-}
