--- realtime/include/cudaq/nvqlink.h
@@ -1,11 +1,0 @@
-/****************************************************************-*- C++ -*-****
- * Copyright (c) 2025 - Present NVIDIA Corporation & Affiliates.               *
- * All rights reserved.                                                        *
- *                                                                             *
- * This source code and the accompanying materials are made available under    *
- * the terms of the Apache License 2.0 which accompanies this distribution.    *
- ******************************************************************************/
-
-#pragma once
-
-#include "nvqlink/nvqlink.h"
