--- realtime/include/cudaq/nvqlink/config.h
@@ -1,15 +1,0 @@
-/****************************************************************-*- C++ -*-****
- * Copyright (c) 2025 - Present NVIDIA Corporation & Affiliates.               *
- * All rights reserved.                                                        *
- *                                                                             *
- * This source code and the accompanying materials are made available under    *
- * the terms of the Apache License 2.0 which accompanies this distribution.    *
- ******************************************************************************/
-
-#pragma once
-
-// Convenience umbrella header - includes all configuration headers
-
-#include "cudaq/nvqlink/daemon/config.h"
-#include "cudaq/nvqlink/network/config.h"
-#include "cudaq/nvqlink/qcs/config.h"
