--- realtime/lib/compiler.cpp
@@ -1,11 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2025 - Present NVIDIA Corporation & Affiliates.               *
- * All rights reserved.                                                        *
- *                                                                             *
- * This source code and the accompanying materials are made available under    *
- * the terms of the Apache License 2.0 which accompanies this distribution.    *
- ******************************************************************************/
-
-#include "cudaq/nvqlink/compiler.h"
-
-INSTANTIATE_REGISTRY_NO_ARGS(cudaq::nvqlink::compiler);
